Move tasks control buttons in page
